IGI/LR2/blogDjangoAndReact/backend_django/requirements.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.4.0
django-countries==6.1.3
djangorestframework==3.11.1
djangorestframework-simplejwt==4.6.0
mysqlclient==2.0.2
Pillow==8.0.1
PyJWT==1.7.1
python-decouple==3.6
pytz==2020.4
sqlparse==0.4.1
